techworldwithnana-gitlab-cicd-demo-app/src/requirements.txt
Flask==2.1.0
py-cpuinfo==7.0.0
psutil==5.8.0
gunicorn==20.1.0
black==20.8b1
flake8==3.9.0
pytest==6.2.5